Moved commands and listeners into directories.
