+++ plugin.py
@@ -5,11 +5,19 @@ from package_control.package_manager import PackageManager
 from package_control.sys_path import add_dependency
 
 from .src.settings import get_settings
-
 from .src.configurations import get_configurations
 
+from .src.commands.build_syntaxes import BuildJsCustomSyntaxesCommand
+from .src.commands.clear_user_data import ClearJsCustomUserDataCommand
+from .src.commands.jsx_close_tag import JsxCloseTagCommand
+from .src.listeners.jsx_close_tag import JsxCloseTagListener
 
-__all__ = ['plugin_loaded', 'plugin_unloaded']
+
+__all__ = [
+    'plugin_loaded', 'plugin_unloaded',
+    'BuildJsCustomSyntaxesCommand', 'ClearJsCustomUserDataCommand', 'JsxCloseTagCommand',
+    'JsxCloseTagListener',
+]
 
 
 PACKAGE_NAME = 'JSCustom'
@@ -23,6 +31,7 @@ def plugin_loaded():
         ensure_dependencies_loaded()
         print('JS Custom: New installation. Building all syntaxes.')
         sublime.active_window().run_command('build_js_custom_syntaxes')
+
     elif events.post_upgrade(PACKAGE_NAME):
         ensure_dependencies_loaded()
         print('JS Custom: Installation upgraded. Rebuilding all syntaxes.')
