+++ src/commands/clear_user_data.py
@@ -1,7 +1,7 @@
 import sublime_plugin
 import shutil
 
-from .src.paths import USER_DATA_PATH
+from ..paths import USER_DATA_PATH
 
 
 class ClearJsCustomUserDataCommand(sublime_plugin.ApplicationCommand):
